requirements.txt
Flask==0.11
flasgger==0.5.13
simplejson==3.8.2
cerberus==1.2
mysql-connector-python==8.0.15
gunicorn